Data Persistence Between Scenes and Sessions
For this commited, I created and implemented data persistence between scenes and play sessions into my project.  The save system tracks the players name with the highest score as well as their score. This data is saved upon quitting the game and restored upon launching the game. The data is serialized and saved in a json file. The player also has the option to delete any currently saved data.